VAN-92: Update TPA Context
