+++ openedx/core/djangoapps/user_authn/utils.py
@@ -5,90 +5,12 @@ Utility functions used during user authentication.
 import random
 import string
 
-import six
 from django.conf import settings
-from django.contrib import messages
 from django.utils import http
-from django.utils.translation import ugettext as _
 from oauth2_provider.models import Application
 from six.moves.urllib.parse import urlparse  # pylint: disable=import-error
 
-from common.djangoapps import third_party_auth
 from openedx.core.djangoapps.site_configuration import helpers as configuration_helpers
-from common.djangoapps.third_party_auth import pipeline
-
-
-def third_party_auth_context(request, redirect_to, tpa_hint=None):
-    """
-    Context for third party auth providers and the currently running pipeline.
-
-    Arguments:
-        request (HttpRequest): The request, used to determine if a pipeline
-            is currently running.
-        redirect_to: The URL to send the user to following successful
-            authentication.
-        tpa_hint (string): An override flag that will return a matching provider
-            as long as its configuration has been enabled
-
-    Returns:
-        dict
-
-    """
-    context = {
-        "currentProvider": None,
-        "providers": [],
-        "secondaryProviders": [],
-        "finishAuthUrl": None,
-        "errorMessage": None,
-        "registerFormSubmitButtonText": _("Create Account"),
-        "syncLearnerProfileData": False,
-        "pipeline_user_details": {}
-    }
-
-    if third_party_auth.is_enabled():
-        for enabled in third_party_auth.provider.Registry.displayed_for_login(tpa_hint=tpa_hint):
-            info = {
-                "id": enabled.provider_id,
-                "name": enabled.name,
-                "iconClass": enabled.icon_class or None,
-                "iconImage": enabled.icon_image.url if enabled.icon_image else None,
-                "loginUrl": pipeline.get_login_url(
-                    enabled.provider_id,
-                    pipeline.AUTH_ENTRY_LOGIN,
-                    redirect_url=redirect_to,
-                ),
-                "registerUrl": pipeline.get_login_url(
-                    enabled.provider_id,
-                    pipeline.AUTH_ENTRY_REGISTER,
-                    redirect_url=redirect_to,
-                ),
-            }
-            context["providers" if not enabled.secondary else "secondaryProviders"].append(info)
-
-        running_pipeline = pipeline.get(request)
-        if running_pipeline is not None:
-            current_provider = third_party_auth.provider.Registry.get_from_pipeline(running_pipeline)
-            user_details = running_pipeline['kwargs']['details']
-            if user_details:
-                context['pipeline_user_details'] = user_details
-
-            if current_provider is not None:
-                context["currentProvider"] = current_provider.name
-                context["finishAuthUrl"] = pipeline.get_complete_url(current_provider.backend_name)
-                context["syncLearnerProfileData"] = current_provider.sync_learner_profile_data
-
-                if current_provider.skip_registration_form:
-                    # As a reliable way of "skipping" the registration form, we just submit it automatically
-                    context["autoSubmitRegForm"] = True
-
-        # Check for any error messages we may want to display:
-        for msg in messages.get_messages(request):
-            if msg.extra_tags.split()[0] == "social-auth":
-                # msg may or may not be translated. Try translating [again] in case we are able to:
-                context["errorMessage"] = _(six.text_type(msg))  # pylint: disable=E7610
-                break
-
-    return context
 
 
 def is_safe_login_or_logout_redirect(redirect_to, request_host, dot_client_id, require_https):
